% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanodrop.R
\name{plot_ndv_spectrum}
\alias{plot_ndv_spectrum}
\title{Plot spectrum of Nanodrop data}
\usage{
plot_ndv_spectrum(ndv_li)
}
\arguments{
\item{ndv_li}{list from read_ndv()}
}
\description{
Plot spectrum of Nanodrop data
}
